fix type conversion warnings  (#17061)
* Fix conversion warning

from unsigned int to float

* Fix type conversion warning

* Fix type conversion warning

* change the return type of getOutlineSize to float

* fix typo

+++ cocos/2d/CCLabel.h
@@ -410,7 +410,7 @@ public:
     /**
     * Return the outline effect size value.
     */
-    int getOutlineSize() const { return _outlineSize; }
+    float getOutlineSize() const { return _outlineSize; }
 
     /**
     * Return current effect type.
